Move MetricsSender into the messaging lib
