+++ sbt_common/messaging/src/main/scala/uk/ac/wellcome/messaging/message/MessageWorker.scala
@@ -1,8 +1,6 @@
 package uk.ac.wellcome.messaging.message
 
 import akka.actor.ActorSystem
-import uk.ac.wellcome.utils.JsonUtil._
-import uk.ac.wellcome.metrics.MetricsSender
 import uk.ac.wellcome.utils.GlobalExecutionContext.context
 
 import scala.concurrent.Future
@@ -10,6 +8,7 @@ import com.twitter.inject.Logging
 
 import scala.concurrent.duration._
 import io.circe.Decoder
+import uk.ac.wellcome.messaging.metrics.MetricsSender
 import uk.ac.wellcome.messaging.sns.NotificationMessage
 import uk.ac.wellcome.messaging.sqs.SQSReader
 import uk.ac.wellcome.utils.JsonUtil._
