+++ sbt_common/messaging/src/main/scala/uk/ac/wellcome/messaging/sqs/SQSWorker.scala
@@ -2,7 +2,7 @@ package uk.ac.wellcome.messaging.sqs
 
 import akka.actor.ActorSystem
 import uk.ac.wellcome.utils.JsonUtil._
-import uk.ac.wellcome.metrics.MetricsSender
+import uk.ac.wellcome.messaging.metrics.MetricsSender
 import uk.ac.wellcome.utils.JsonUtil.fromJson
 import uk.ac.wellcome.utils.GlobalExecutionContext.context
 import scala.concurrent.Future
