+++ sbt_common/messaging/src/main/scala/uk/ac/wellcome/messaging/sqs/SQSWorkerToDynamo.scala
@@ -4,7 +4,7 @@ import akka.actor.ActorSystem
 import com.amazonaws.services.dynamodbv2.model.ConditionalCheckFailedException
 import io.circe.Decoder
 import uk.ac.wellcome.exceptions.GracefulFailureException
-import uk.ac.wellcome.metrics.MetricsSender
+import uk.ac.wellcome.messaging.metrics.MetricsSender
 import uk.ac.wellcome.utils.GlobalExecutionContext.context
 import uk.ac.wellcome.utils.JsonUtil._
