+++ sbt_common/messaging/src/test/scala/uk/ac/wellcome/messaging/metrics/MetricsSenderTest.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.metrics
+package uk.ac.wellcome.messaging.metrics
 
 import java.time.Instant
 import java.time.temporal.ChronoUnit
