+++ sierra_adapter/sierra_item_merger/src/main/scala/uk/ac/wellcome/platform/sierra_item_merger/services/SierraItemMergerWorkerService.scala
@@ -3,8 +3,8 @@ package uk.ac.wellcome.platform.sierra_item_merger.services
 import akka.actor.ActorSystem
 import com.google.inject.Inject
 import io.circe.Decoder
+import uk.ac.wellcome.messaging.metrics.MetricsSender
 import uk.ac.wellcome.messaging.sqs.{SQSReader, SQSWorker, SQSWorkerToDynamo}
-import uk.ac.wellcome.metrics.MetricsSender
 import uk.ac.wellcome.models.transformable.sierra.SierraItemRecord
 import uk.ac.wellcome.utils.GlobalExecutionContext.context
 import uk.ac.wellcome.utils.JsonUtil._
